refactor: common/auto_close.ts と common/vc_reservation/ を feat-recruit 直下に移動
